Extensive changes based on Jeremy Hylton's investigations.  These should considerably help the performance of the archiver. Specifically:
ArchiveMail(): Create a lock file (and lock it), just after the fork.
Jeremy observes that there is a race condition when many posts show up
in a short amount of time.  By creating a lock file we make sure that
the separate archiver processes won't clobber each other.

Use the new LockFile module.

Move the (c)StringIO import to the top of the file.