Robustness changes.  Specifically:
When processing the archive, lay claim to the archiver lock so that
any posts that come in while the archiver is running will get
blocked.  This isn't optimal because we don't really know how long the
archiver will run (I give it 3 hours max by default).
HyperArchive.processUnixMailbox() should actually refresh() the lock
while it's working.

Also, the MailList object is locked too, although the lifetime is
shorter.  I wonder if this is still necessary?

Use getopt module to process options.  Added a docstring which becomes
the usage message.