New pipeline delivery module
"""Deliver a message via `sendmail' drop-off.

This module delivers the message via the command line interface to the
sendmail program.  It should work for sendmail clones like Postfix.  It is
expected that sendmail handles final delivery, message queueing, etc.  The
recipient list is only trivially split so that the command line is less that
4k in size.

"""