HELDMSG, SUBSCRIPTION: removed these from here since they're only used locally in ListAdmin.py
ADMINDB_PAGE_TEXT_LIMIT: new variable which controls how big an
excerpt to include in the admindb text field.  Current default is 4k.
This number is passed directly to fileobj.read(), so a negative number
means put the whole message in the field (which could slow down the
web page).