Held posts are now stored in a separate file in the file system, instead of marshalled into the requests.db file.  The file name will always be $PREFIX/data/heldmsg-<listname>-<id> where <id> is the request id integer value.
HoldMessage(): Store the message text in the heldmsg file as described
above.  In the requests.db marshal, we just include the filename in
the last slot.  Although this is redundant information, it makes
calculating the file name a little easier (hmm, maybe we should remove
this anyway).

__handlepost(): Watch out for the message file being removed, and
raise a LostHeldMessage in that situation.  In any case, once the
message is dispatched, be sure to unlink the message file
(transforming expected exceptions into LostHeldMessage).