AcknowlegePosts => AcknowledgePosts
Convert message delivery to pipeline architecture by using
HandlerAPI.DeliverToUser() for fast track delivery of the
acknowledgement to the sender.   We need to be a little tricky digging
the original sender out of a msg object attribute, since CookHeaders
(which sets the attribute) changes it for the outgoing message.  If
msg.original_sender isn't found, use the normal msg.GetSender().