process() => DeliverToList()
DeliverToUser(): New function which implements a pipeline for direct
delivery to a specified user.  This pipeline is used for all admin
notifications, and other things like user post acks.  It sets
msg.fastrack attribute to 1 so other modules can know which delivery
track this message is on.

This lets us use the same module for all delivery, or separate ones
for delivery to list and delivery to user.