A serious rewrite to make it more robust, and to add a somewhat useful command line interface.  This latter allows you to manually poll a specific list's newsgroup for a range of message numbers, sending them to the list.  Using this interface does /not/ update the watermark file.
We also don't update the watermark entry for a specific list until
that list's child process (which does the actually gating of the
messages from Usenet to mailing list) exits with a zero status.  If it
exits with non-zero, it is taken as a complete failure of the gating
process and the watermark isn't updated.  This could cause
duplicates, so a really anal approach would be to create a pipe
between child and parent and write the new watermark after each
message is successfully gated.