process(): RFC 1894 allows for multiple addresses in a single bounce notification, so we need to handle this.
Sadly, MTAs that support DSN use a bunch of different headers which
might be more useful in picking out the actual list member.
Final-recipient is required but may not correspond to our list
member.  Original-recipient, which /would/ be a list member, is
optional.  Some MTAs seem to use X-Actual-Recipient to point to our
list member, but others use that for the final recip.  :(