NotExplicitlyAllowed: New class to represent message hold conditions described below.
process(): When members_posting_only is not set, but posters /is/ set,
the semantics are that only those explicitly listed in posters can
post without admin approval.  Even though I think this is AFU'd, at
least this patch fixes things so it works the same as it did in
Mailman 1.1.