Update(): Call NewVars(), a new function.
NewVars(): Set the various attributes for the autoresponder if they
don't exist yet.  Even though the comments say this shouldn't be
necessary, because InitTempVars() should take care of it, I don't
think that really works anymore.  NewVars() is necessary so that the
new mixin class attributes get initialized on existing mailing lists
which are restored from disk after the code gets updated.