The handler module which actually does the auto-responding.  A few notes:
If the message contains an "X-Ack: No" header, then it is never
auto-responded to.  This, in fact, is used to break both internal
recursion and potential mail loops.

If the message contains an "X-Ack: Yes" header, then it is
auto-responded to even if the grace period hasn't expired (but the
grace period is updated).