InitVars(): added autorespond_requests, autoresponse_request_text, and request_responses.
GetConfigInfo(): I figured it /was/ worth adding an autoresponder on
the -request address after all.  This works the same as the -admin,
and postings autoresponder with one variation: you now have three
choices

1. no autoreply
2. autoreply, but discard the original message
3. autoreply, and forward the message off to the mail command
   processing robot

This is useful for lists that essentially to shut off the command
handler, but it may be overkill.  It's definitely useful to autoreply
to the -request address after all.

Also, I slightly reformated the HTML for this page, putting the common
stuff up in the preamble.