New feature (hopefully, the last one before feature freeze for 1.2 :)
Elaborate the Reply-To munging to handle parallel lists, e.g. a
checkins list where discussions about changes should be conducted on a
parallel development list.  Specifically,

There's a new configuration attribute called `reply_to_address' which
contains the string address to stuff in the Reply-To: header.  The
config attr reply_goes_to_list now takes three values

    0 (no change) means no reply-to munging occurs
    1 (no change) reply-to is munged to the list address
    2 (new value) reply-to is munged to reply_to_address

In both cases, if the original message contains a Reply-To: header, it
is discarded.  This could be controversial; it might be better to
honor an existing Reply-To: header in case #2, but that seems like it
would be less useful.

GetConfigInfo(): Add the necessary machinery to effect the above
change.  Also, use TEXTFIELDWIDTH as the standard width for all string
entry fields.