Conversion of Mailman 1.1 pending admin database to 1.2/2.0 external requests.db database.  This should be the last major hurdle for the new version.  Specifically,
NewRequestsDatabase(): This knows about the 1.1 style mlist.requests
dictionary, which served as the pending requests database.  It sucks
everything out of this dict and resubmits the requests using the new
APIs.  Worst thing that happens its that the timestamps get blown
away.  Oh well.

older(): Removed since it's not used anywhere.