Implementation of file queing on total or partial delivery failures. Specifically,
process(): move smptlib.SMTP() into the try/except so that
socket.errors on the connection can be flagged and dealt with as total
delivery failures.

Better error logging.

If the connection goes through but smtplib throws an
SMTPRecipientsRefused exception, then Only persistent transient
failures are queued, i.e. those with error code < 500.  For failed
recipients with codes >= 500, these are permanent failures and we
do a bounce registration instead.

queue_message(): handles all the actual queuing of files for future
re-delivery attempts.  Two files per message are written to
$prefix/qfiles: one is a .txt containing the full text of the message,
the other is a .db file containing a marshal of this information:

    listname - Python string naming the list for delivery
    recips   - Python list of addr strings
    attepts  - number of total delivery attempts for this queued message
    last_recip_count - number of recips the last time this delivery
		       was re-tried (not sure how useful that will
		       be).

The base name of the two files is the SHA hexdigest hash of the
message text.