Added an update to move gate_watermark entries into the list-specific usenet_watermark attribute.  We get rid of the gate_watermark file once we've updated.
Consistency in imports.

Try to determine whether the templates/options.html warning needs to
be printed or not.  It's only printed if we can't determine the last
Mailman version and we are not running a fresh install (i.e. there are
entries in the logs subdir).

Write the current Mailman version (HEX_VERSION) as a hex
representation string into the file data/last_mailman_version.  This
is so the /next/ update can determine what version we're updating
from.  Lays the groundwork for future updates.