Substantial changes for robustification of deliveries.
dispose_message(): Compatibility function for delivery of pre-2.0beta3
version qfiles and 2.0b3 version qfiles (whose marshaled state now
contains a schema version number).

open_list(): Since qrunner can attempt delivery against the same list
multiple times, cache the list name and the MailList object in a
global, and reuse it if possible.  The only thing that has to be
managed explicitly is acquiring and relinquishing the lock.  That fine
though 'cause that's what we actually want to do!

main(): Integrate with the new delivery stuff; use the msgdata
dictionary for message metadata.  Also, dig out the 'kids' info, which
is a list of subproc pids to wait on.  Re-queue the message if
delivery did not successfully complete.