Elaborations related to the new delivery stuff.  Specifically,
HoldMessage(): Take an optional msgdata dictionary, which gets stored
along with all the other information related to this held message in
the held message database.

__handlepost(): Add some backwards compatibility code for pre 2.0b3
held message databases now that the held record can be 6 elements long
(this should really not be a length-dependent tuple).  Also, use
HandlerAPI.DeliverToList() to send approved messages out, and check
the return value to see if the message needs to be enqueued.