Fixes inspired or provided by Harald Meland.  Specifically,
lock(): It is an expected condition that the global linkfile can have
a linkcount of 1.  This can arise due to a race condition between
proc1 that is in the middle of unlocking, and proc2 which is trying to
acquire the lock.  In this case, we simply log the condition and try
again later.

Also, if we're already locked, we should not unlink our tmpfname.

unlock(): We don't need to wrap self.locked() in a try/except.

__break(): Updated the comment at the head of the method based on
Harald's observation.  Touching the lockfile doesn't eliminate the
race condition, just makes it more unlikely.

Also, we don't need to wrap the self.__read() in a try/except.

Finally, elaborate the unit test suit to use /dev/random if it exists
(i.e. on Linux).