This script now does most of the work that post, mailcmd, and mailowner used to do.  Those scripts now just quickly queue the messages into qfiles/.  This was the only way to make sure that we never hit MTAs delivery command time limits.
QRUNNER_LOCK_LIFETIME: moved to Defaults.py

dispose_message(): Knows how to dispose of messages queued by post,
mailcmd, or mailowner.  mailcmd sets the `torequest' msgdata key, and
mailowner sets the `toadmin' msgdata key.

For torequest messages, we send the message through ParseMailCommands
and always return 0.

For toadmin messages, we send the message through the bounce scanner,
and if that fails to find a bounce, we send the message on to the list
owner (with all the proper queuing on errors).

For all other messages, we do nothing different.

main(): Be a little more careful about closing the file descriptors.