HandleRequest(): Takes three new additional arguments, `preserve', `forward', and `addr' which are passed in from admindb.py.  These are only useful for held postings, and are only passed on to __handlepost().
preserve is a flag that says whether the held message should be copied
to the `spam' directory or not.

forward is a flag which says that the message should be forwarded on
to an addition email address, which is contained in the addr
parameter.  Both most be set for forwarding to occur.

Finally, HandleRequest() only deletes the database entry if the status
of the sub-method is true.  This is to support the new `deferred' flag
for held postings.

__handlepost(): Takes the three new parameters as described above.
Uses shutil to do the `preserve' copy, and HandlerAPI.DeliverToUser()
to do the forwarding (after inserting a `Resent-To' header).