Several UI and functional changes.  Better layout and better text in places.  For held subscriptions, the displayed order of `subscribe' and `refuse' is changed so that the refused button is right next to the refusal explanation.
After suggestions by many people, split the headers and body into
different textboxes for held messages.  Before held messages often got
truncated in their headers, and the bodies couldn't be seen.

Instead of providing a button to view the entire message (which would
be more difficult to implement through the admindb authentication
blockade), it is now possible to forward the message to a given
address.  This defaults to the <listname>-admin address.  Use this
combined with the new `Defer' action to get a copy of the full message
for review before deciding what to do with it.

In addition to the new `Defer' action, held messages can be
`preserved' for the site admin by copying the message to the
$PREFIX/spam directory.  This is useful for abusive messages that need
closer or long term scrutiny.

Message rejection notices are taken from the
msgdata['rejection-notice'] value, placed their by the Hold.py
module.