Get rid of the -s/--stderrs and --quiet/-q options.  Use syslog() for all logging activities except stderr output from child tracebacks. Get rid of unused imports.
poll_newsgroup(): Reorganized so as not to use the crufty
QuickEscape.  Also, be sure to update mlist.usenet_watermark even if
we see a message that's already been through the list.  When
delivering the message to the list, enqueue it immediately (by adding
the _enqueue_immediate flag to msgdata) instead of trying to deliver
it.  This lets qrunner deliver the message as appropriate.

reap(): Move to Utils.py

process_lists(): Don't try to be fancy about catching exceptions in
the child.  Just let them percolate up and Python will do the right
thing by logging it and exiting with a non-zero status.

main(): Use standard LockFile module for acquiring the gate_news lock.