Use syslog() interface instead of writing to stderr.  All syslog errors go to the logs/gate_news file, while any exceptions that percolate to the top would go to logs/error.
poll_newsgroup(): Always update the watermark, even if the message has
been seen before (i.e. we found an X-BeenThere: header) or if an
nntp.error_temp occurred (which can happen if the message has
expired).  Any other exceptions will not be caught and thus will not
update the watermark.

process_lists(): Be extra paranoid about the forks and child
processes.  First of all, the fork could fail and if so, we stop
cruising through the mailing lists and return to the parent the
dictionary of child pids gathered so far.  Second, make absolutely
sure that the child process exits without returning control to the
caller.  If we complete successfully or if a TimeOutError occurred, we
exit with status 0, otherwise we print a traceback and exit with
status 1.

main(): First off, we added a keepalive function to geep the gate_news
lock fresh while we're waiting for our child processes to exit.
Second, we added a failsafe around process_list() -- which does a fork
-- so that in the unlikely situation where the child process could
have gotten returned here, we exit immediately.