DumbBTree.__init__(): Rewrote the file opening and unmarshaling code to be more robust so that if either fail, we end up with an empty self.dict and self.sorted.
Note that the archiver subprocess will still fail with an exception.
Fixing this will require much more work on the archiver as a whole,
and isn't worth it right now.  But this fix averts the problem when
regenerating the archive from scratch using bin/arch, so at least
corrupt archives can be rebuilt.