ArchiveMail(): First, extended the signature to take a msgdata dictionary, which gets filled in with the child pid.  This allows it to be reaped by functions higher up in the call chain.
Second, and more importantly, be quite paranoid about the fork and
child processes.  Make absolutely sure that the child process exits.
Wrap all the child code in try blocks so that no matter what, the lock
will get released (by calling unlock() with the unconditionally flag
set).  If no exception occurs, the child will exit with status 0,
otherwise the traceback will be printed to stderr, a message
indicating which list has a corrupt archive is logged to logs/error,
and the child process will exit with status 1.

Extend the archive.lock to 1 hour.