Hackish support for `volatile' properties in the web interface. I.e. these are selections that can be presented in an admin page, but don't directly affect the value of a saved list attribute.  This is used for a "Mass Catchup" button in the gateway category.  Volatile property convention is that the name starts with an underscore. Specifically,
GetItemGuiValue(): For Radios and Toggles, only do a getattr() of
varname if the name does not start with an underscore.  Otherwise
default the checked value to 0.  This is bogus because it would be
better to take the default value from the property description, but
that's too much of an API change for now.

ChangeOptions(): If the property name starts with an underscore, do
something different.  This has to check for specific volatile
parameters, of which the only one currently is `_mass_catchup'.  If
that's the property and the value is true, then set usenet_watermark
to None, which is the signal to gate_news to do a mass catchup
(i.e. set the watermark to the last article currently on the server).

Note that the GatewayManager category used to always do a mass catchup
when gateway_to_mail was toggled.  This isn't always the right thing,
so this change allows the admin to explicitly do a catch up.