dispose_message(): A refinement of the message disposition scheme. For messages that have toadmin set (i.e. came to the -admin address), we only do bounce detection before sending off to the list owners. Specifically, we don't do bounce-loop detection.
All messages sent to the owners will get Errors-To: set to the -owner
address.  It is the -owner address that does bounce-loop detection
before sending off to the list owners.

The reason this is done is so that bounces arriving at the -admin
address but not matching any detectors will be forwarded on the list
owners.  But bogus list owner address must be dealt with as well, so
these are handled by the -owner address.

This is all documented at the top of the file.