One more try...
Enqueue(): Only calculate a new filebase (SHA hash) if the object
wasn't assigned one in its constructor.  Go back to including the
current float time in the hash input.

Message.__init__(): Take an optional `filebase' argument which can be
restored from an enqueued message.  This guarantees that a message,
once assigned a hash value for its filebase, continues to use that
filebase for its entire life.