RegisterBounce(): Remove the saveifdirty flag from the argument list, and unwrap from the try/finally saving the mailing list.  We'll assume the caller always saves the list.
HandleBouncingAddress(): Don't do a msg.rewindbody()/msg.fp.read() to
get the text of the message.  When called from qrunner, the file
handle will be closed, and this will cause a fatal error in
SMTPDirect, which in turn causes multiple sends until all offending
email addresses are disabled.  The right way to do this now is to use
the msg.body attribute.  This closes SF bugs #111843 and #111176.