FormatOptionHelp(), GetItemGuiDescr(): Several changes which improve the usability of the web interface, closing SF bug #114126, Jitterbug PR#148.  Specifically,
- There is now a submit button on the details page, and you can use
  that page to change options.  This should work better now that
  details by default don't pop open a new window, however this is a
  warning at the bottom of the page about synchronization issues.

- Details pages now include the standard footer too.

- Volatile config variables (e.g. _mass_catchup) have a note saying
  that setting the value performs an action, but doesn't change
  permanent state.

ChangeOptions(): The above change uncovered what I believe is a bug,
where if the form data is missing a property found on the category's
options, it would be skipped /unless/ the type of the property were
Text, String, or EmailList.  I don't understand why that special case
was there, so I changed this logic to always skip the property if it
isn't found in the form data.  Otherwise submitting from a details
page could cause some confusing behavior.

GetConfigOptions(): I don't think the caching of mlist.GetConfigInfo()
is worth it.  The lookups should be plenty fast.