Several changes in support of David Champion's SF patch #101331. Specifically,
MAILMAN_UID and MAILMAN_GID are now available in mm_cfg, so use these
as the gid to check group-ownership against, and as the user to re-run
the script as if there are errors.

checkall(): Now knows about all of PREFIX, EXEC_PREFIX, and VAR_PREFIX
and checks the permissions of them all.  It's smart about knowing when
any of these directories are the same.