First step in the conversion to a multiple-queue model.  This script gets simplified considerably because the bulk of the logic is moved into classes in the Mailman.Queue package.  The next step might be to make the queues long running processes watchdogged by the qrunner.
QRUNNERS is a list of classes to instantiate to run through the
various queue directories.  Currently there are only incoming,
outgoing, and usenet queues.  Each qrunner is run in it's own child
process (via fork), so we need to be more careful about race
conditions in the re-queuing logic.

main(): Pretty simple now, it just forks and waits on the child
processes.