Enqueue(): Allow for queuing to alternative queue directories.  If newdata or kws has an item called "_whichq", this specifies the directory to queue to, defaulting to INQUEUE_DIR.
The metadata "filebase" is always set to __filebase, not just when the
metadata is created for the first time.

When writing the .db file (i.e. containing the metadata), always write
it after the .msg text file, and always write it to a temporary file,
with an os.rename() shuffle.  This should avoid race conditions when
queuing a message to another queue because the queue runners only key
off of files ending in ".db".

Requeue(): New method which calls Enqueue() after zapping
self.__filebase so it'll be recalculated.