Latest round of jcrey's I18N patches.  Specifically,
Mark various strings as translatable.

HoldSubscription(): New parameter `lang' which is the user's preferred
language.  This is stored in the pending database.  Messages to the
administrator are sent in the list's preferred language, while
messages back to the user are sent in the user's preferred language.

__handlesubscription(): Pull the user's preferred language from the
held record and pass that to __refuse().

__refuse(): Pass the selected language to maketext().