Switchboard class which supports enqueuing and dequeuing messages from the various qfiles/* directories.  A Switchboard object must be constructed with the directory that it is to watch over (along with hash space slicing information).
Switchboard also supports returning the list of files in its hash
space slice, via the files() directory.  Also, there are `protected'
methods _ext_write() and _ext_read() which handle the low-level
reading and writing of the message metadata.  In the base class, these
are unimplemented, and subclasses override to provide the raw file
format for the metadata.