__init__(): Accept information for hash space slicing (slice number and total number of slices).
_dispose(): Do the work of flattening the message and posting it to
the news server.  We really don't need a pipeline here.

prepare_message(): This gets the message ready for posting via news.
It's debatable whether the prep work on the message should be done
here or in the ToUsenet module.  We do it here to avoid having to make
a copy of the message in ToUsenet.

Convert to mimelib.