Next round of big i18n patches.
main(): Do two step creation and locking of mailing list so we can
have a better error message when we couldn't acquire the list lock.
Also, make sure that once we have the lock, all work is wrapped in a
try/finally so that the list won't be left in a locked state on
error.