More i18n changes.
De-string-module-ification.

Fix some _() wrappings to use local variables.

At global module scope, set the language to the server's default,
until we know which mailing list this request is being made on.

main(): Set the document's language to the server's default,
until we know which mailing list this request is being made on.  Then,
once we know the list, set the system and document languages to the
list's preferred, until we know the user making the request.  Finally,
once we've got the user, set the system and document languages to the
user's preferred.

(get rid of os.environ['LANG'] settings)