More i18n changes.
De-string-module-ification and some Python 2.0-isms.

Fix some _() wrappings to use local variables.

At global module scope, set the language to the server's default,
until we know which mailing list this request is being made on.

main(): Set the document's language to the server's default,
until we know which mailing list this request is being made on.  Then,
once we know the list, set the system and document languages to the
list's preferred, until we know the user making the request.

(get rid of os.environ['LANG'] settings)

process_form(): NEW SUPPORT -- if the user has subscribed without
filling in either the password or the confirmed password, Mailman will
automatically assign a password for them.  If either field is filled
but not both, and error message is still displayed.

PrintResults() -> print_results()