More i18n changes.
At module scope, set the system language to the server's default,
until we know which list this is a request on.

main(): Don't lock the list until absolutely necessary.  Also, once
the list is instantiated, set the system language to the list's
preferred language.  Once the html results document object is created,
set its language to the list's preferred language too.

HandleRequests() -> process_form()
PrintRequests() -> show_requests()

handle_no_list(): Be sure to set the document's language to the
server's default (since there is no instantiated list at this point).

show_requests(): Support DEFER and DISCARD on subscription requests,
just like as for held posts.

PrintAddMemberRequest(): Inlined and removed.

PrintPostRequest() -> show_post_requests()

show_post_requests(): Convert to using the mimelib interface (it's a
minor shame that we have to reparse the message object to display it
on this page).  Watch for mimelib.Errors.MessageParseError's which
could result if the held message is corrupted.  If this occurs, we
discard the message, but that may not be the right thing to do.

process_form(): Don't save the MailList object here; that happens in
the finally: clause in main().

Also some minor verbiage and formatting changes.